// packages/botonic-nlu/src/types.d.ts
import { Tensor } from '@tensorflow/tfjs-node';

/* Word Embeddings */
export type WordEmbeddingType = '10k-fasttext' | 'glove';
export type WordEmbeddingDimension = 50 | 300;
export interface WordEmbeddingsConfig {
  type: WordEmbeddingType;
  dimension: WordEmbeddingDimension;
  language: Language;
  vocabulary: Vocabulary;
}

/* Data preprocessing */
export interface Normalizer {
  normalize(text: string): string;
}

export interface Stemmer {
  stem(text: string, language: Language): string;
}

export interface Tokenizer {
  tokenize(text: string): string[];
}

/* Sets */
export type DataSet = { label: string; feature: string }[];
export type InputSet = Tensor;
export type OutputSet = Tensor;

/* Intents codification */
export type IntentEncoder = { [intent: string]: number };
export type IntentDecoder = { [id: number]: string };
export type EncodedPrediction = { intentId: number; confidence: number }[];
export type DecodedPrediction = { intent: string; confidence: number }[];

/* Model Creation */
export interface ModelData {
  language: Language;
  intents: IntentDecoder;
  maxSeqLen: number;
  vocabulary: Vocabulary;
}

export interface ModelParameters {
  maxSeqLen: number;
  learningRate: number;
  intentsCount: number;
  trainableEmbeddings: boolean;
}

/* Model Training */
export interface TrainingParameters {
  X: InputSet;
  y: OutputSet;
  epochs: number;
  batchSize: number;
  validationSplit: number;
}

/* Language Data */
export type Vocabulary = { [word: string]: number };
export type Language =
  | 'aa'
  | 'ab'
  | 'ae'
  | 'af'
  | 'ak'
  | 'am'
  | 'an'
  | 'ar'
  | 'as'
  | 'av'
  | 'ay'
  | 'az'
  | 'ba'
  | 'be'
  | 'bg'
  | 'bh'
  | 'bi'
  | 'bm'
  | 'bn'
  | 'bo'
  | 'br'
  | 'bs'
  | 'ca'
  | 'ce'
  | 'ch'
  | 'co'
  | 'cr'
  | 'cs'
  | 'cu'
  | 'cv'
  | 'cy'
  | 'da'
  | 'de'
  | 'dv'
  | 'dz'
  | 'ee'
  | 'el'
  | 'en'
  | 'eo'
  | 'es'
  | 'et'
  | 'eu'
  | 'fa'
  | 'ff'
  | 'fi'
  | 'fj'
  | 'fo'
  | 'fr'
  | 'fy'
  | 'ga'
  | 'gd'
  | 'gl'
  | 'gn'
  | 'gu'
  | 'gv'
  | 'ha'
  | 'he'
  | 'hi'
  | 'ho'
  | 'hr'
  | 'ht'
  | 'hu'
  | 'hy'
  | 'hz'
  | 'ia'
  | 'id'
  | 'ie'
  | 'ig'
  | 'ii'
  | 'ik'
  | 'io'
  | 'is'
  | 'it'
  | 'iu'
  | 'ja'
  | 'jv'
  | 'ka'
  | 'kg'
  | 'ki'
  | 'kj'
  | 'kk'
  | 'kl'
  | 'km'
  | 'kn'
  | 'ko'
  | 'kr'
  | 'ks'
  | 'ku'
  | 'kv'
  | 'kw'
  | 'ky'
  | 'la'
  | 'lb'
  | 'lg'
  | 'li'
  | 'ln'
  | 'lo'
  | 'lt'
  | 'lu'
  | 'lv'
  | 'mg'
  | 'mh'
  | 'mi'
  | 'mk'
  | 'ml'
  | 'mn'
  | 'mr'
  | 'ms'
  | 'mt'
  | 'my'
  | 'na'
  | 'nb'
  | 'nd'
  | 'ne'
  | 'ng'
  | 'nl'
  | 'nn'
  | 'no'
  | 'nr'
  | 'nv'
  | 'ny'
  | 'oc'
  | 'oj'
  | 'om'
  | 'or'
  | 'os'
  | 'pa'
  | 'pi'
  | 'pl'
  | 'ps'
  | 'pt'
  | 'qu'
  | 'rm'
  | 'rn'
  | 'ro'
  | 'ru'
  | 'rw'
  | 'sa'
  | 'sc'
  | 'sd'
  | 'se'
  | 'sg'
  | 'si'
  | 'sk'
  | 'sl'
  | 'sm'
  | 'sn'
  | 'so'
  | 'sq'
  | 'sr'
  | 'ss'
  | 'st'
  | 'su'
  | 'sv'
  | 'sw'
  | 'ta'
  | 'te'
  | 'tg'
  | 'th'
  | 'ti'
  | 'tk'
  | 'tl'
  | 'tn'
  | 'to'
  | 'tr'
  | 'ts'
  | 'tt'
  | 'tw'
  | 'ty'
  | 'ug'
  | 'uk'
  | 'ur'
  | 'uz'
  | 've'
  | 'vi'
  | 'vo'
  | 'wa'
  | 'wo'
  | 'xh'
  | 'yi'
  | 'yo'
  | 'za'
  | 'zh'
  | 'zu';